+++ publish/CHANGELOG.md
@@ -1,5 +1,6 @@
 - v1.31
   - Adds blueprint rotation based on the center piece.
+  - Fixes blueprint scale not being used.
 
 - v1.30
   - Adds new fields `centerX` and `centerY` to the `expand_world.yaml` (replaces `curveX` and `curveY`).